# Conflicts:
#	AvatarApp/Avatar.App.Api/appsettings.json
#	AvatarApp/Avatar.App.Infrastructure/Migrations/20200523205328_AddFireBaseId.Designer.cs
#	AvatarApp/Avatar.App.Infrastructure/Migrations/20200523205328_AddFireBaseId.cs
#	AvatarApp/Avatar.App.Infrastructure/Migrations/20200918134115_ProdMigration.Designer.cs
#	AvatarApp/Avatar.App.Infrastructure/Migrations/AvatarAppContextModelSnapshot.cs

+++ AvatarApp/Avatar.App.Infrastructure/EfBaseRepository.cs
@@ -24,12 +24,12 @@ namespace Avatar.App.Infrastructure
             StorageService = storageService;
         }
 
-        public T GetById(int id)
+        public T GetById(long id)
         {
             return DbContext.Set<T>().Find(id);
         }
 
-        public async Task<T> GetByIdAsync(int id)
+        public async Task<T> GetByIdAsync(long id)
         {
             return await DbContext.Set<T>().FindAsync(id);
         }
